requirements.txt
streamlit
pandas
psycopg2-binary
fpdf
python-dotenv

